Update Alpine GLIBC packages to 2.28-r0
The Alpine GLIBC packages haven't been updated in a while, so this
change updates to the latest available release in Sasha Gerrand's
repository. The GPG public key URL has changed and is no longer
packaged along with GLIBC releases.

It's easier to read variables when they're not declared across
several lines, so I've decided to relax the maximum line length to
120 characters.

Finally, this change explicitly invokes privilege escalation in order
to install the GLIBC packages. Since the CI process has been using
the `root` user inside the official Alpine Docker image(s), this
was easy to miss.