Add Ansible Galaxy downloads badge
Re-arranges the existing badges, renames the link aliases with either
`link-` or `badge-` as appropriate because it's easier to read/sort,
and adds a badge showing how many times this role has been downloaded
from Ansible Galaxy. The badges are converted to "flat-square" style
and the license badge is now dynamically rendered (just in case it
changes in the future).