Fix permission errors when checking SDKMAN install for different user
When SDKMAN is installed for a different user than the one Ansible is
connecting with for the first time, it is likely that the user home
directory permissions do not allow traversal for the Ansible user.

To fix the problem, let's switch to the root user for this check,
which should work in virtually every reasonable case.